requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
altair==5.1.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.3.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==1.10.0
beautifulsoup4==4.12.2
black==23.10.1
bleach==6.0.0
blessed==1.20.0
cachetools==5.3.1
cchardet==2.1.7
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chex==0.1.7
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.2
coloredlogs==15.0.1
colorful==0.5.5
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
dm-env==1.6
dm-meltingpot==2.2.0
dm-tree==0.1.8
dmlab2d==1.0.0
EMD-signal==1.5.1
exceptiongroup==1.1.3
executing==2.1.0
fastapi==0.103.1
fastjsonschema==2.18.0
filelock==3.12.2
flatbuffers==24.3.25
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.0
gast==0.4.0
gevent==23.9.0.post1
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
gpustat==1.1.1
greenlet==2.0.2
grpcio==1.57.0
gym==0.23.1
gym-notices==0.0.8
Gymnasium==0.26.3
gymnasium-notices==0.0.1
h11==0.14.0
h5py==3.12.1
humanfriendly==10.0
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
immutabledict==4.2.0
imutils==0.5.4
ipykernel==6.25.1
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jax==0.4.14
jaxlib==0.4.14
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==3.6.0
kiwisolver==1.4.5
kubernetes==27.2.0
lazy_loader==0.3
libclang==16.0.6
lit==16.0.6
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mat73==0.62
matplotlib==3.7.2
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mdurl==0.1.2
mistune==3.0.1
ml-collections==0.1.1
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
neurokit2==0.2.5
nltk==3.8.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.108
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.1
onnxruntime==1.15.1
opencensus==0.11.2
opencensus-context==0.1.3
opencv-python==4.8.0.76
opentelemetry-api==1.1.0
opentelemetry-exporter-otlp==1.1.0
opentelemetry-exporter-otlp-proto-grpc==1.1.0
opentelemetry-proto==1.1.0
opentelemetry-sdk==1.1.0
opentelemetry-semantic-conventions==0.20b0
opt-einsum==3.3.0
optree==0.13.1
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pox==0.3.3
ppft==1.7.6.7
prometheus-client==0.13.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
pyarrow==7.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.12
pyeeg==0.0.2
pygame==2.6.1
Pygments==2.16.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
ray==2.2.0
ray-cpp==2.2.0
reactivex==4.0.4
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.17.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
tensorflow-probability==0.24.0
termcolor==2.3.0
terminado==0.17.1
tf_keras==2.17.0
threadpoolctl==3.2.0
tifffile==2023.8.30
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.64.1
traitlets==5.9.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.0
uvicorn==0.23.2
virtualenv==20.24.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.6.2
Werkzeug==2.3.7
widgetsnbextension==4.0.8
wrapt==1.15.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
